correction, have replaced removeClientFromChannel() by deleteChannelClient()

+++ srcs/Channel.cpp
@@ -66,7 +66,7 @@ void	Channel::addChannelClient(Client *client)
 {
 	std::vector<Client*>::iterator it;
 
-	if (channelClients.size() >= _userLimit)
+	if (channelClients.size() >= static_cast<size_t>(_userLimit))
 	{
 		client->reply(ERR_CHANNELISFULL(client->getNickname(), _name));
 		return ;
